@@ -19,101 +19,62 @@ log_level = "DEBUG"
 filterwarnings = [
     "ignore: The output driver does not natively support DateTime type",
     "ignore: Field .* create as date field, though DateTime requested",
+    "ignore: Field .* created as String field, though DateTime requested",
     "ignore: Geometry is in a geographic CRS",
     "ignore: Normalized/laundered field name",
     "ignore: Non-conformant content for record 0 in column",
     "ignore: Non-conformant content for record 1 in column",
+    "ignore: Using GFO_IO_ENGINE=fiona is deprecated",
 ]
 
 [tool.ruff]
 line-length = 88
-target-version = "py39"
-extend-exclude = ["docs/*", "local_ignore/*"]
+target-version = "py310"
+extend-exclude = ["docs/conf.py", "local_ignore/*"]
 
 [tool.ruff.lint]
 select = [
-    # pyflakes
-    "F",
-    # pycodestyle
-    "E",
-    "W",
-    # pyupgrade
-    "UP",
-    # flake8-bugbear
-    "B",
-    # flake8-debugger
-    "T10",
-    # flake8-simplify
-    # "SIM",
-    # pylint
-    "PLC",
-    "PLE",
-    "PLR",
-    "PLW",
-    # misc lints
-    "PIE",
-    # implicit string concatenation
-    "ISC",
-    # type-checking imports
-    "TCH",
-    # comprehensions
-    "C4",
-    # Ruff-specific rules
-    "RUF",
-    # isort
-    "I",
-    # pydocstyle
-    "D",
+    "YTT",  # flake8-2020
+    "ANN",  # flake8-annotations
+    "B",  # flake8-bugbear
+    "A",  # flake8-builtins
+    "C4",  # flake8-comprehensions
+    "T10",  # flake8-debugger
+    "ISC",  # flake8-implicit string concatenation
+    # "G",  # flake8-logging-format
+    "PIE",  # flake8-pie misc lints
+    # "SIM",  # flake8-simplify
+    "TC",  # flake8-type-checking imports
+    "ARG",  # flake8-unused-arguments
+    "PTH",  # flake8-use-pathlib
+    "FLY",  # flynt
+    "I",  # isort
+    "NPY",  # NumPy-specific rules
+    "E",  # pycodestyle errors
+    "W",  # pycodestyle warnings
+    "D",  # pydocstyle
+    "F",  # pyflakes
+    "PLC",  # pylint convention
+    "PLE",  # pylint error
+    "PLR",  # pylint refactor
+    "PLW",  # pylint warning
+    "UP",  # pyupgrade
+    "RUF",  # Ruff-specific rules
 ]
 
 ignore = [
     ### Intentionally disabled
-    # module level import not at top of file
-    "E402",
-    # do not assign a lambda expression, use a def
-    "E731",
-    # mutable-argument-default
-    "B006",
-    # unused-loop-control-variable
-    "B007",
-    # get-attr-with-constant
-    "B009",
-    # Only works with python >=3.10
-    "B905",
-    # dict literals
-    "C408",
-    # Too many arguments to function call
-    "PLR0913",
-    # Too many returns
-    "PLR0911",
-    # Too many branches
-    "PLR0912",
-    # Too many statements
-    "PLR0915",
-    # Magic number
-    "PLR2004",
-    # Redefined loop name
-    "PLW2901",
-    # Global statements are discouraged
-    "PLW0603",
-    # compare-to-empty-string
-    "PLC1901",
-
-    ### Additional checks that don't pass yet
-    # Useless statement
-    "B018",
-    # Within an except clause, raise exceptions with ...
-    "B904",
-    # Consider `elif` instead of `else` then `if` to remove indentation level
-    "PLR5501",
-    # collection-literal-concatenation
-    "RUF005",
-    # Mutable class attributes should be annotated with `typing.ClassVar`,
-    "RUF012"
+    "PLR0913",  # Too many arguments to function call
+    "PLR0911",  # Too many returns
+    "PLR0912",  # Too many branches
+    "PLR0915",  # Too many statements
+    "PLR2004",  # Magic number
+    "PLW2901",  # loop variable overwritten by assignment target
 ]
 
 [tool.ruff.lint.per-file-ignores]
-"tests/*" = ["D"]
+"tests/*" = ["ANN", "D"]
+"perftests/*" = ["ANN", "D"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "google"
