@@ -6,7 +6,6 @@ import inspect
 import logging
 import tempfile
 from pathlib import Path
-from typing import Optional
 
 import pandas as pd
 
@@ -33,8 +32,8 @@ class RunResult:
         operation: str,
         operation_descr: str,
         secs_taken: float,
-        run_details: Optional[dict] = None,
-    ):
+        run_details: dict | None = None,
+    ) -> None:
         """Constructor for a RunResult.
 
         Args:
@@ -54,15 +53,15 @@ class RunResult:
         self.secs_taken = secs_taken
         self.run_details = run_details
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         """Format the result."""
         return f"{self.__class__}({self.__dict__})"
 
 
 def run_benchmarks(
-    modules_to_run: Optional[list[str]] = None,
-    functions_to_run: Optional[list[str]] = None,
-):
+    modules_to_run: list[str] | None = None,
+    functions_to_run: list[str] | None = None,
+) -> None:
     """Run all benchmarks specified.
 
     Args:
