@@ -4,19 +4,20 @@ import datetime
 import logging
 import os
 from collections.abc import Iterable
-from typing import Any, Optional
+from types import TracebackType
+from typing import Any
 
 logger = logging.getLogger(__name__)
 
 
-class MissingRuntimeDependencyError(Exception):
-    """Exception raised when an unsupported SQL statement is passed.
+class MissingRuntimeDependencyError(RuntimeError):
+    """Exception raised when a geofileops runtime dependency is missing.
 
     Attributes:
         message (str): Exception message
     """
 
-    def __init__(self, message):
+    def __init__(self, message: str) -> None:
         self.message = message
         super().__init__(self.message)
 
@@ -80,9 +81,9 @@ def report_progress(
     start_time: datetime.datetime,
     nb_done: int,
     nb_todo: int,
-    operation: Optional[str] = None,
+    operation: str | None = None,
     nb_parallel: int = 1,
-):
+) -> None:
     # If logging level not enabled for INFO, no progress reporting...
     if logger.isEnabledFor(logging.INFO) is False:
         return
@@ -104,9 +105,9 @@ def format_progress(
     start_time: datetime.datetime,
     nb_done: int,
     nb_todo: int,
-    operation: Optional[str] = None,
+    operation: str | None = None,
     nb_parallel: int = 1,
-) -> Optional[str]:
+) -> str | None:
     # Init
     time_passed = (datetime.datetime.now() - start_time).total_seconds()
     pct_progress = 100.0 - (nb_todo - nb_done) * 100 / nb_todo
@@ -145,9 +146,9 @@ def format_progress(
         return message
 
 
-def formatbytes(bytes: float):
+def formatbytes(nb_bytes: float) -> str:
     """Return the given bytes as a human friendly KB, MB, GB, or TB string."""
-    bytes_float = float(bytes)
+    bytes_float = float(nb_bytes)
     KB = float(1024)
     MB = float(KB**2)  # 1,048,576
     GB = float(KB**3)  # 1,073,741,824
@@ -161,16 +162,16 @@ def formatbytes(bytes: float):
         return f"{bytes_float / MB:.2f} MB"
     elif GB <= bytes_float < TB:
         return f"{bytes_float / GB:.2f} GB"
-    elif TB <= bytes_float:
+    else:
         return f"{bytes_float / TB:.2f} TB"
 
 
 def prepare_for_serialize(data: dict) -> dict:
     prepared: dict[str, Any] = {}
     for key, value in data.items():
-        if isinstance(value, (dict)):
+        if isinstance(value, dict):
             prepared[key] = prepare_for_serialize(value)
-        elif isinstance(value, (list, tuple)):
+        elif isinstance(value, list | tuple):
             prepared[key] = value
         else:
             prepared[key] = str(value)
@@ -184,15 +185,17 @@ class TempEnv:
     Existing values for variables are backed up and reset when the scope is left,
     variables that didn't exist before are deleted again.
 
+    If value is None, the environment variable is deleted within the context.
+
     Args:
         envs (Dict[str, Any]): dict with environment variables to set.
     """
 
-    def __init__(self, envs: dict[str, Any]):
+    def __init__(self, envs: dict[str, Any]) -> None:
         self._envs_backup: dict[str, str] = {}
         self._envs = envs
 
-    def __enter__(self):
+    def __enter__(self) -> None:
         # Only if a name and value is specified...
         for name, value in self._envs.items():
             # If the environment variable is already defined, make backup
@@ -200,15 +203,18 @@ class TempEnv:
                 self._envs_backup[name] = os.environ[name]
 
             # Set env variable to value
-            os.environ[name] = str(value)
+            if value is None:
+                if name in os.environ:
+                    del os.environ[name]
+            else:
+                os.environ[name] = str(value)
 
-    def __exit__(self, type, value, traceback):
+    def __exit__(self, type: type, value: Exception, traceback: TracebackType) -> None:  # noqa: A002
         # Set variables that were backed up back to original value
         for name, env_value in self._envs_backup.items():
             # Recover backed up value
             os.environ[name] = env_value
         # For variables without backup, remove them
         for name, _ in self._envs.items():
-            if name not in self._envs_backup:
-                if name in os.environ:
-                    del os.environ[name]
+            if name not in self._envs_backup and name in os.environ:
+                del os.environ[name]
