@@ -1,6 +1,5 @@
 # import datetime
 import logging
-from typing import Optional
 
 import pygeoops
 import shapely
@@ -18,7 +17,7 @@ def gfo_difference_collection(
     geom_to_subtract_wkb: bytes,
     keep_geom_type: int = 0,
     subdivide_coords: int = 2000,
-) -> Optional[bytes]:
+) -> bytes | None:
     """Applies the difference of geom_to_subtract on geom.
 
     If the input geometry has many points, they can be subdivided in smaller parts
@@ -52,8 +51,6 @@ def gfo_difference_collection(
         if geom_wkb is None:
             return None
         if geom_to_subtract_wkb is None:
-            return geom_wkb
-        if subdivide_coords <= 0:
             return geom_wkb
 
         # Extract wkb's, and return if empty
@@ -97,7 +94,7 @@ def gfo_difference_collection(
         return None
 
 
-def gfo_reduceprecision(geom_wkb: bytes, gridsize: int) -> Optional[bytes]:
+def gfo_reduceprecision(geom_wkb: bytes, gridsize: int) -> bytes | None:
     """Reduces the precision of the geometry to the gridsize specified.
 
     If reducing the precison leads to a topologyerror, retries after applying make_valid
@@ -161,7 +158,7 @@ def gfo_reduceprecision(geom_wkb: bytes, gridsize: int) -> Optional[bytes]:
 def gfo_split(
     geom_wkb: bytes,
     blade_wkb: bytes,
-) -> Optional[bytes]:
+) -> bytes | None:
     """Applies a split in the geom using the blade specified.
 
     Args:
@@ -217,7 +214,7 @@ def gfo_split(
         return None
 
 
-def gfo_subdivide(geom_wkb: bytes, coords: int = 2000):
+def gfo_subdivide(geom_wkb: bytes, coords: int = 2000) -> bytes | None:
     """Divide the input geometry to smaller parts using rectilinear lines.
 
     Args:
@@ -378,10 +375,10 @@ def _int2bool(value: int, variable_name: str) -> bool:
         raise TypeError(
             f"{variable_name} must be int (0: False or 1: True), not {type(value)}"
         )
-    if value not in [0, 1]:
-        raise ValueError(f"{variable_name} has invalid value (0=False/1=True): {value}")
 
-    if value == 0:
+    if value == 1:
+        return True
+    elif value == 0:
         return False
     else:
-        return True
+        raise ValueError(f"{variable_name} has invalid value (0=False/1=True): {value}")
