@@ -3,20 +3,25 @@
 from benchmark import benchmarker
 
 
-def main():
+def main() -> None:
     """Run the benchmark function(s)."""
     functions_to_run = [
-        "export_by_location_intersects_complexpoly",
-        # "export_by_location_intersects",
-        # "join_by_location_intersects",
+        # "buffer_spatialite",
         # "clip",
+        # "clip_agri_complexpoly",
+        "dissolve_nogroupby",
+        "dissolve_groupby",
+        # "export_by_location_intersects_complexpoly",
+        # "export_by_location_intersects",
         # "intersection",
         # "intersection_complexpoly_agri",
+        # "intersection_complexpoly_complexpoly",
         # "intersection_gridsize",
+        # "join_by_location_intersects",
         # "symmetric_difference_complexpolys_agri",
         # "union",
     ]
-    # Run all bechmark functions
+    # Run all benchmark functions
     # functions_to_run = None
     benchmarker.run_benchmarks(["benchmarks_geofileops"], functions_to_run)
 
