@@ -3,6 +3,7 @@
 import ast
 import csv
 import enum
+import warnings
 from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Union
@@ -10,7 +11,9 @@ from typing import TYPE_CHECKING, Any, Union
 from osgeo import gdal
 from osgeo_utils.auxiliary.util import GetOutputDriversFor
 
-if TYPE_CHECKING:
+from geofileops.util._geopath_util import GeoPath
+
+if TYPE_CHECKING:  # pragma: no cover
     import os
 
 # Enable exceptions for GDAL
@@ -34,9 +37,9 @@ class GeofileTypeInfo:
 geofiletypes: dict[str, GeofileTypeInfo] = {}
 
 
-def _init_geofiletypes():
+def _init_geofiletypes() -> None:
     geofiletypes_path = Path(__file__).resolve().parent / "geofiletypes.csv"
-    with open(geofiletypes_path) as file:
+    with geofiletypes_path.open() as file:
         # Set skipinitialspace to True so the csv can be formatted for readability
         csv.register_dialect("geofiletype_dialect", skipinitialspace=True, strict=True)
         reader = csv.DictReader(file, dialect="geofiletype_dialect")
@@ -79,7 +82,7 @@ class GeofileType(enum.Enum):
     FlatGeobuf = enum.auto()
 
     @classmethod
-    def _missing_(cls, value):
+    def _missing_(cls, value: object) -> Union["GeofileType", None]:
         """Expand options in the GeofileType() constructor.
 
         Args:
@@ -93,16 +96,16 @@ class GeofileType(enum.Enum):
             [GeofileType]: The corresponding GeometryType.
         """
 
-        def get_geofiletype_for_suffix(suffix: str):
+        def get_geofiletype_for_suffix(suffix: str) -> GeofileType:
             suffix_lower = suffix.lower()
-            for geofiletype in geofiletypes:
-                if suffix_lower in geofiletypes[geofiletype].suffixes:
+            for geofiletype, geofiletype_info in geofiletypes.items():
+                if suffix_lower in geofiletype_info.suffixes:
                     return GeofileType[geofiletype]
             raise ValueError(f"Unknown extension {suffix}")
 
-        def get_geofiletype_for_ogrdriver(ogrdriver: str):
-            for geofiletype in geofiletypes:
-                driver = geofiletypes[geofiletype].ogrdriver
+        def get_geofiletype_for_ogrdriver(ogrdriver: str) -> GeofileType:
+            for geofiletype, geofiletype_info in geofiletypes.items():
+                driver = geofiletype_info.ogrdriver
                 if driver is not None and driver == ogrdriver:
                     return GeofileType[geofiletype]
             raise ValueError(f"Unknown ogr driver {ogrdriver}")
@@ -149,10 +152,7 @@ class GeofileType(enum.Enum):
     @property
     def is_singlelayer(self) -> bool:
         """Returns True if a file of this GeofileType can only have one layer."""
-        if self.is_spatialite_based:
-            return False
-        else:
-            return True
+        return not self.is_spatialite_based
 
 
 # Init!
@@ -166,7 +166,7 @@ class GeofileInfo:
         driver (str): the relevant gdal driver for the file.
     """
 
-    def __init__(self, path: Path):
+    def __init__(self, path: Union[str, "os.PathLike[Any]"]) -> None:
         """Constructor of Layerinfo.
 
         Args:
@@ -176,7 +176,7 @@ class GeofileInfo:
         self.driver = get_driver(path=path)
         self.geofiletype_info = geofiletypes.get(self.driver.replace(" ", ""))
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         """Overrides the representation property of GeofileInfo."""
         return f"{self.__class__}({self.__dict__})"
 
@@ -210,7 +210,7 @@ class GeofileInfo:
 
     @property
     def default_spatial_index(self) -> bool:
-        """Returns True if this geofile can only have one layer."""
+        """Returns True if this file type gets a spatial index by default."""
         if self.geofiletype_info is not None:
             return self.geofiletype_info.default_spatial_index
         else:
@@ -229,32 +229,54 @@ def get_driver(path: Union[str, "os.PathLike[Any]"]) -> str:
     """Get the gdal driver name for the file specified.
 
     Args:
-        path (PathLike): The file path.
+        path (PathLike): The file path. |GDAL_vsi| paths are also supported.
 
     Returns:
         str: The OGR driver name.
-    """
-    path = Path(path)
 
-    def get_driver_for_path(input_path) -> str:
+    .. |GDAL_vsi| raw:: html
+
+        <a href="https://gdal.org/en/stable/user/virtual_file_systems.html" target="_blank">GDAL vsi</a>
+
+    """  # noqa: E501
+    # gdal.OpenEx is relatively slow on windows, so for straightforward cases, avoid it.
+    suffix = GeoPath(path).suffix_full.lower()
+    if suffix in (".gpkg", ".gpkg.zip"):
+        return "GPKG"
+    elif suffix in (".shp", ".shp.zip"):
+        return "ESRI Shapefile"
+
+    def get_driver_for_path(
+        input_path: Union[str, "os.PathLike[Any]"], driver_prefix: str | None
+    ) -> str:
         # If there is no suffix, possibly it is only a suffix, so prefix with filename
-        if input_path.suffix == "":
+        local_path = input_path
+        if Path(input_path).suffix == "":
             local_path = f"temp{input_path}"
-        else:
-            local_path = input_path
 
         drivers = GetOutputDriversFor(local_path, is_raster=False)
         if len(drivers) == 1:
             return drivers[0]
-        else:
+        elif len(drivers) == 0:
             raise ValueError(
-                "Could not infer driver from path. Please specify driver explicitly by "
-                "prefixing the file path with '<DRIVER>:', e.g. 'GPKG:path'. "
+                "Could not infer driver from path. You can try to specify the driver "
+                "by prefixing the file path with '<DRIVER>:', e.g. 'GPKG:path'. "
                 f"Path: {input_path}"
             )
+        else:
+            if driver_prefix is not None and driver_prefix in drivers:
+                return driver_prefix
 
-    # If the file exists, determine the driver based on the file.
-    datasource = None
+            warnings.warn(
+                f"Multiple drivers found, using first one of: {drivers}. If you want "
+                "another driver, you can try to specify the driver by prefixing the "
+                f"file path with '<DRIVER>:', e.g. 'GPKG:path'. Path: {input_path}",
+                UserWarning,
+                stacklevel=2,
+            )
+            return drivers[0]
+
+    # Try to determine the driver by opening the file.
     try:
         datasource = gdal.OpenEx(
             str(path), nOpenFlags=gdal.OF_VECTOR | gdal.OF_READONLY | gdal.OF_SHARED
@@ -262,9 +284,28 @@ def get_driver(path: Union[str, "os.PathLike[Any]"]) -> str:
         driver = datasource.GetDriver()
         drivername = driver.ShortName
     except Exception as ex:
-        try:
-            drivername = get_driver_for_path(path)
-        except Exception:
+        ex_str = str(ex).lower()
+        driver_prefix_list = str(path).split(":", 1)
+        driver_prefix = (
+            driver_prefix_list[0]
+            if len(driver_prefix_list) > 0 and len(driver_prefix_list[0]) > 1
+            else None
+        )
+        if (
+            "no such file or directory" in ex_str
+            or "not recognized as being in a supported file format" in ex_str
+            or "not recognized as a supported file format" in ex_str
+            or driver_prefix is not None
+        ):
+            # If the file does not exist or if, for some cases like a csv file,
+            # it is e.g. an empty file that was not recognized yet, try to get the
+            # driver based on only the path.
+            try:
+                drivername = get_driver_for_path(path, driver_prefix)
+            except Exception:
+                ex.args = (f"get_driver error for {path}: {ex}",)
+                raise
+        else:
             ex.args = (f"get_driver error for {path}: {ex}",)
             raise
     finally:
@@ -285,4 +326,4 @@ def get_geofileinfo(path: Union[str, "os.PathLike[Any]"]) -> GeofileInfo:
     Returns:
         GeofileInfo: _description_
     """
-    return GeofileInfo(path=Path(path))
+    return GeofileInfo(path=path)
